@@ -1,9 +1,11 @@
 {
   "name": "ex5",
   "dependencies": {
+    "axios": "^1.1.2",
+    "cors": "^2.8.5",
     "express": "^4.18.1",
-    "react": "^15.5.4",
-    "react-dom": "^15.5.4"
+    "react": "^18.3.0-next-aa9988e5e-20221010",
+    "react-dom": "^18.3.0-next-aa9988e5e-20221010"
   },
   "version": "0.1.0",
   "devDependencies": {
